ai_resume_selector/requirements.txt
endee
openai
gradio
pdfplumber
python-docx
python-dotenv
sentence-transformers
